scraper/requirements.txt
pdfplumber>=0.10.0
scikit-learn>=1.3.0
nltk>=3.8.0
python-dotenv>=1.0.0
rich>=13.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pymarkdownlnt>=0.9.0
mdformat>=0.7.0
